+++ transcriptic/objects.py
@@ -636,3 +636,4 @@ class Container(_BaseObject):
 
         """
         return "Container(%s)" % (str(self.name))
+