+++ transcriptic/routes.py
@@ -52,6 +52,10 @@ def launch_protocol(api_root, org_id, protocol_id):
     return "{api_root}/{org_id}/protocols/{protocol_id}/launch".format(**locals())
 
 
+def get_launch_request(api_root, org_id, protocol_id, launch_request_id):
+    return "{api_root}/{org_id}/protocols/{protocol_id}/launch/{launch_request_id}".format(**locals())
+
+
 def post_release(api_root, org_id, package_id):
     return "{api_root}/{org_id}/packages/{package_id}/releases/".format(**locals())