+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 ## `transcriptic` Changelog
 
 ## Unreleased
+## v3.7.0
 Added
 - Ability to browse your inventory using the `transcriptic inventory` command E.g. `transcriptic inventory water`
 - Ability to launch protocols remotely using the `--remote` flag. E.g. `transcriptic launch Pipetting --remote`
 - Ability to view available remote protocols for launching using `transcriptic protocols --remote`
+- Ability for `transcriptic summarize` to retrieve resource strings with the `--lookup` flag
 
 Fixed
 - resources route has been updated to match web return